Allow old ipykernel with newer traitlets/ipython
Since 4.0.1 ipykernel uses ipython_genutils however this dependency was
not made explicit in ipykernel's setup.py until ipykernel 5.5.x.
Instead the dependency was at the time satisfied by pulling in ipython
or traitlets which both depended on ipython_genutils.

Now, ipython and traitlets have dropped that dependency so when
installing an old ipykernel with a newer ipython/traitlets, a
ModuleNotFoundError is reported:

```
[...]
  File "/usr/share/miniconda/envs/test/lib/python3.7/site-packages/ipykernel/__init__.py", line 2, in <module>
    from .connect import *
  File "/usr/share/miniconda/envs/test/lib/python3.7/site-packages/ipykernel/connect.py", line 13, in <module>
    from ipython_genutils.path import filefind
ModuleNotFoundError: No module named 'ipython_genutils'
```

This happens e.g. with ipykernel 5.5.5, ipython 5.8.0, traitlets 5.4.0.
This combination arises when installing and old sagelib 9.1.

Here, we fix this by making the dependency on ipython_genutils explicit
for all versions of ipykernel that use it. Note that ipython_genutils
has only two versions on conda-forge 0.1.0 and 0.2.0 so it should be
safe to always use the latest (as ipykernel did at the time.)
ipython_genutils is not developed anymore so there should not be any
breaking changes in the future.