Fix pexpect
Closes conda-forge/pexpect-feedstock#26

Sometimes when installing a package for py38 or earlier (when non-noarch
is an option), the required dependency
ptyprocess is missing. Example:
<https://github.com/conda-forge/poetry-feedstock/blob/cc7b7d853791182668e2277ac4f93beaff63cb79/recipe/meta.yaml>

This also corrects cases where the ptyprocess pin ">=0.5" was missing.

pexpect has had the pin ptyprocess >=0.5 since pexpect v4.0. The
ptyprocess package used to not exist on Windows, so pexpect required a
"# [not win]" selector, but at some point it
was made noarch. This will allow "pip check" to pass on Windows. To
avoid disturbing our ancestors, we avoid patching below py36.
In particular:

- Unpatched: py27, py34, py35, py36
- Patched: py37, py38
- No change: noarch

Expected impact is low since there are no direct or transitive
dependencies other than "ptyprocess >=0.5", and both pexpect and
ptyprocess are now noarch.