Patch old packages to require the old SWIG runtime
The SWIG runtime version recently changed from 4 to 5,
and packages that share SWIG-wrapped classes need to use
the same SWIG runtime to work correctly. Thanks to
conda-forge/swig-feedstock#58 new builds of such packages
will depend on v5 of the swig-abi package. This patch
adds a dependency on v4 of swig-abi for older versions
of two such affected packages (`imp` and `rmf).

Note that v3 of the SWIG runtime was way back in 2008,
long before any of the packages here were built, so we
can safely tag all old packages as v4.